requirements.txt
fastapi
uvicorn
loguru
pyjwt
python-multipart
email-validator
sqlalchemy==1.4.0b3
aiomysql
aioredis
redislite
msgpack